lib: move paperclip dependency to http-server
The paperclip crate is used to generate the OpenAPI specs.
It's better to have the dependency in the http-server because it's the
only place where it's really needed, plus that way, we can define the
structs in the library without about worrying the limitations of
deriving the "Apiv2schema" trait from paperclip.