Dockerfile: actually use Alpine for building instead of cross-compiling to musl
This is way simpler and more reliable. The problem at this point in time
is that we are pulling in OpenSSL as a dependency, and this requires a
build of this library with musl. Rather than rebuilding OpenSSL
manually, let's just compile the project in a musl-based distro... and
why not use Alpine, which we already use for running the service!